Move unix_path_to_win into conda.common.path
